backend/indexer/requirements.txt
spacy
langchain
pandas
qdrant-client
tiktoken
google-cloud-secret-manager==2.10.0
google-cloud-storage
sentence-transformers